+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - A valid external link requires at least one legal character after the protocol
 - Attribute delimiter of table captions and table cells
+- Unmatched closing tag
 
 ## v2.1.9